+++ R/class-smoothing_options.R
@@ -9,11 +9,11 @@ setClass("smoothing_options",
     boundary_method = "character"
   ),
   prototype = list(
-    order_poly = 1,
+    order_poly = 3,
     season = NA_real_,
     kernel_fun = "epanechnikov",
     bwidth = NA_real_,
-    boundary_method = "extend"
+    boundary_method = "shorten"
   )
 )
 
@@ -42,7 +42,7 @@ setValidity("smoothing_options", function(object) {
 #'@import methods
 #' 
 #'@param order_poly the order of the local polynomials used for estimating the
-#' smooth nonparametric trend; the default is \code{1}.
+#' smooth nonparametric trend; the default is \code{3}.
 #'@param season the frequency of observations per time unit, for example per 
 #'year; set to \code{12} for monthly data and to \code{4} for quarterly data
 #'and so on; the default is \code{NA_real_}, which leads to an automated 
@@ -61,9 +61,9 @@ setValidity("smoothing_options", function(object) {
 #'triggers a data-driven selection of an globally optimal bandwidth when
 #'the output of this function is passed to a smoothing function.
 #'@param boundary_method a single character value; it indicates, what bandwidth 
-#'method to use at boundary points; for \code{"extend"}, the default, the 
+#'method to use at boundary points; for \code{"extend"}, the 
 #'smoothing window around boundary points will be extended towards the center of 
-#'the data; for \code{"shorten"}, the window width will keep decreasing at 
+#'the data; for \code{"shorten"}, the default, the window width will keep decreasing at 
 #'boundary points when approaching the very first and the very last observation.
 #'
 #'@export
@@ -102,11 +102,11 @@ setValidity("smoothing_options", function(object) {
 #'}
 
 set_options <- function(
-    order_poly = 1,
+    order_poly = 3,
     season = NA_real_,
     kernel_fun = "epanechnikov",
     bwidth = NA_real_,
-    boundary_method = "extend") {
+    boundary_method = "shorten") {
   
       methods::new("smoothing_options",
         order_poly = order_poly,