+++ src/RcppExports.cpp
@@ -161,8 +161,8 @@ BEGIN_RCPP
 END_RCPP
 }
 // BICarmaCpp
-double BICarmaCpp(const arma::vec& Xt, const int p, const int q, const int armamean);
-RcppExport SEXP _deseats_BICarmaCpp(SEXP XtSEXP, SEXP pSEXP, SEXP qSEXP, SEXP armameanSEXP) {
+double BICarmaCpp(const arma::vec& Xt, const int p, const int q, const int armamean, const int inf_c);
+RcppExport SEXP _deseats_BICarmaCpp(SEXP XtSEXP, SEXP pSEXP, SEXP qSEXP, SEXP armameanSEXP, SEXP inf_cSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
     Rcpp::RNGScope rcpp_rngScope_gen;
@@ -170,13 +170,14 @@ BEGIN_RCPP
     Rcpp::traits::input_parameter< const int >::type p(pSEXP);
     Rcpp::traits::input_parameter< const int >::type q(qSEXP);
     Rcpp::traits::input_parameter< const int >::type armamean(armameanSEXP);
-    rcpp_result_gen = Rcpp::wrap(BICarmaCpp(Xt, p, q, armamean));
+    Rcpp::traits::input_parameter< const int >::type inf_c(inf_cSEXP);
+    rcpp_result_gen = Rcpp::wrap(BICarmaCpp(Xt, p, q, armamean, inf_c));
     return rcpp_result_gen;
 END_RCPP
 }
 // selectOrderBIC
-Rcpp::NumericVector selectOrderBIC(const arma::vec& Xt, const int pmin, const int pmax, const int qmin, const int qmax, const int armamean);
-RcppExport SEXP _deseats_selectOrderBIC(SEXP XtSEXP, SEXP pminSEXP, SEXP pmaxSEXP, SEXP qminSEXP, SEXP qmaxSEXP, SEXP armameanSEXP) {
+Rcpp::NumericVector selectOrderBIC(const arma::vec& Xt, const int pmin, const int pmax, const int qmin, const int qmax, const int armamean, const int inf_c);
+RcppExport SEXP _deseats_selectOrderBIC(SEXP XtSEXP, SEXP pminSEXP, SEXP pmaxSEXP, SEXP qminSEXP, SEXP qmaxSEXP, SEXP armameanSEXP, SEXP inf_cSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
     Rcpp::RNGScope rcpp_rngScope_gen;
@@ -186,13 +187,14 @@ BEGIN_RCPP
     Rcpp::traits::input_parameter< const int >::type qmin(qminSEXP);
     Rcpp::traits::input_parameter< const int >::type qmax(qmaxSEXP);
     Rcpp::traits::input_parameter< const int >::type armamean(armameanSEXP);
-    rcpp_result_gen = Rcpp::wrap(selectOrderBIC(Xt, pmin, pmax, qmin, qmax, armamean));
+    Rcpp::traits::input_parameter< const int >::type inf_c(inf_cSEXP);
+    rcpp_result_gen = Rcpp::wrap(selectOrderBIC(Xt, pmin, pmax, qmin, qmax, armamean, inf_c));
     return rcpp_result_gen;
 END_RCPP
 }
 // armaoptCpp
-Rcpp::List armaoptCpp(const arma::vec& Xt, const int narmin, const int narmax, const int nmamin, const int nmamax, const int armamean);
-RcppExport SEXP _deseats_armaoptCpp(SEXP XtSEXP, SEXP narminSEXP, SEXP narmaxSEXP, SEXP nmaminSEXP, SEXP nmamaxSEXP, SEXP armameanSEXP) {
+Rcpp::List armaoptCpp(const arma::vec& Xt, const int narmin, const int narmax, const int nmamin, const int nmamax, const int armamean, const int inf_c);
+RcppExport SEXP _deseats_armaoptCpp(SEXP XtSEXP, SEXP narminSEXP, SEXP narmaxSEXP, SEXP nmaminSEXP, SEXP nmamaxSEXP, SEXP armameanSEXP, SEXP inf_cSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
     Rcpp::RNGScope rcpp_rngScope_gen;
@@ -202,7 +204,8 @@ BEGIN_RCPP
     Rcpp::traits::input_parameter< const int >::type nmamin(nmaminSEXP);
     Rcpp::traits::input_parameter< const int >::type nmamax(nmamaxSEXP);
     Rcpp::traits::input_parameter< const int >::type armamean(armameanSEXP);
-    rcpp_result_gen = Rcpp::wrap(armaoptCpp(Xt, narmin, narmax, nmamin, nmamax, armamean));
+    Rcpp::traits::input_parameter< const int >::type inf_c(inf_cSEXP);
+    rcpp_result_gen = Rcpp::wrap(armaoptCpp(Xt, narmin, narmax, nmamin, nmamax, armamean, inf_c));
     return rcpp_result_gen;
 END_RCPP
 }
@@ -290,8 +293,8 @@ BEGIN_RCPP
 END_RCPP
 }
 // algorithmCpp
-Rcpp::List algorithmCpp(const arma::vec& yt, const int p, const int s, const int mu, const double bStart, const int CF, const int errors, const double cb, const int bb, const int errm, const double expo, const int narmin, const int narmax, const int nmamin, const int nmamax, const int armamean);
-RcppExport SEXP _deseats_algorithmCpp(SEXP ytSEXP, SEXP pSEXP, SEXP sSEXP, SEXP muSEXP, SEXP bStartSEXP, SEXP CFSEXP, SEXP errorsSEXP, SEXP cbSEXP, SEXP bbSEXP, SEXP errmSEXP, SEXP expoSEXP, SEXP narminSEXP, SEXP narmaxSEXP, SEXP nmaminSEXP, SEXP nmamaxSEXP, SEXP armameanSEXP) {
+Rcpp::List algorithmCpp(const arma::vec& yt, const int p, const int s, const int mu, const double bStart, const int CF, const int errors, const double cb, const int bb, const int errm, const double expo, const int narmin, const int narmax, const int nmamin, const int nmamax, const int armamean, const int inf_c);
+RcppExport SEXP _deseats_algorithmCpp(SEXP ytSEXP, SEXP pSEXP, SEXP sSEXP, SEXP muSEXP, SEXP bStartSEXP, SEXP CFSEXP, SEXP errorsSEXP, SEXP cbSEXP, SEXP bbSEXP, SEXP errmSEXP, SEXP expoSEXP, SEXP narminSEXP, SEXP narmaxSEXP, SEXP nmaminSEXP, SEXP nmamaxSEXP, SEXP armameanSEXP, SEXP inf_cSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
     Rcpp::RNGScope rcpp_rngScope_gen;
@@ -311,7 +314,8 @@ BEGIN_RCPP
     Rcpp::traits::input_parameter< const int >::type nmamin(nmaminSEXP);
     Rcpp::traits::input_parameter< const int >::type nmamax(nmamaxSEXP);
     Rcpp::traits::input_parameter< const int >::type armamean(armameanSEXP);
-    rcpp_result_gen = Rcpp::wrap(algorithmCpp(yt, p, s, mu, bStart, CF, errors, cb, bb, errm, expo, narmin, narmax, nmamin, nmamax, armamean));
+    Rcpp::traits::input_parameter< const int >::type inf_c(inf_cSEXP);
+    rcpp_result_gen = Rcpp::wrap(algorithmCpp(yt, p, s, mu, bStart, CF, errors, cb, bb, errm, expo, narmin, narmax, nmamin, nmamax, armamean, inf_c));
     return rcpp_result_gen;
 END_RCPP
 }
@@ -372,16 +376,16 @@ static const R_CallMethodDef CallEntries[] = {
     {"_deseats_rseqCpp", (DL_FUNC) &_deseats_rseqCpp, 2},
     {"_deseats_factorialCpp", (DL_FUNC) &_deseats_factorialCpp, 1},
     {"_deseats_armaCpp", (DL_FUNC) &_deseats_armaCpp, 4},
-    {"_deseats_BICarmaCpp", (DL_FUNC) &_deseats_BICarmaCpp, 4},
-    {"_deseats_selectOrderBIC", (DL_FUNC) &_deseats_selectOrderBIC, 6},
-    {"_deseats_armaoptCpp", (DL_FUNC) &_deseats_armaoptCpp, 6},
+    {"_deseats_BICarmaCpp", (DL_FUNC) &_deseats_BICarmaCpp, 5},
+    {"_deseats_selectOrderBIC", (DL_FUNC) &_deseats_selectOrderBIC, 7},
+    {"_deseats_armaoptCpp", (DL_FUNC) &_deseats_armaoptCpp, 7},
     {"_deseats_arma_sumacovCpp", (DL_FUNC) &_deseats_arma_sumacovCpp, 1},
     {"_deseats_acovCpp", (DL_FUNC) &_deseats_acovCpp, 2},
     {"_deseats_cf0Cpp", (DL_FUNC) &_deseats_cf0Cpp, 1},
     {"_deseats_fitteddeseatsCpp", (DL_FUNC) &_deseats_fitteddeseatsCpp, 6},
     {"_deseats_residdeseatsCpp", (DL_FUNC) &_deseats_residdeseatsCpp, 6},
     {"_deseats_derivdeseatsCpp", (DL_FUNC) &_deseats_derivdeseatsCpp, 7},
-    {"_deseats_algorithmCpp", (DL_FUNC) &_deseats_algorithmCpp, 16},
+    {"_deseats_algorithmCpp", (DL_FUNC) &_deseats_algorithmCpp, 17},
     {"_deseats_wls_Cpp", (DL_FUNC) &_deseats_wls_Cpp, 3},
     {"_deseats_find_sub_Cpp", (DL_FUNC) &_deseats_find_sub_Cpp, 4},
     {"_deseats_llin_calc_Cpp", (DL_FUNC) &_deseats_llin_calc_Cpp, 7},