+++ tests/testthat/test-Cpp_arma_related_functions.R
@@ -6,42 +6,42 @@ p <- 1
 q <- 0
 fit1_a <- armaCpp(Xt, p = p, q = q, armamean = TRUE)
 fit1_b <- stats::arima(Xt, order = c(p, 0, q))
-BIC1_a <- BICarmaCpp(Xt, p = p, q = q, armamean = TRUE)
+BIC1_a <- BICarmaCpp(Xt, p = p, q = q, armamean = TRUE, inf_c = 1)
 BIC1_b <- -2 * fit1_b$loglik + ln * (p + q)
 
 p <- 0
 q <- 1
 fit2_a <- armaCpp(Xt, p = p, q = q, armamean = TRUE)
 fit2_b <- stats::arima(Xt, order = c(p, 0, q))
-BIC2_a <- BICarmaCpp(Xt, p = p, q = q, armamean = TRUE)
+BIC2_a <- BICarmaCpp(Xt, p = p, q = q, armamean = TRUE, inf_c = 1)
 BIC2_b <- -2 * fit2_b$loglik + ln * (p + q)
 
 p <- 0
 q <- 0
 fit3_a <- armaCpp(Xt, p = p, q = q, armamean = TRUE)
 fit3_b <- stats::arima(Xt, order = c(p, 0, q))
-BIC3_a <- BICarmaCpp(Xt, p = p, q = q, armamean = TRUE)
+BIC3_a <- BICarmaCpp(Xt, p = p, q = q, armamean = TRUE, inf_c = 1)
 BIC3_b <- -2 * fit3_b$loglik + ln * (p + q)
 
 p <- 2
 q <- 0
 fit4_a <- armaCpp(Xt, p = p, q = q, armamean = TRUE)
 fit4_b <- stats::arima(Xt, order = c(p, 0, q))
-BIC4_a <- BICarmaCpp(Xt, p = p, q = q, armamean = TRUE)
+BIC4_a <- BICarmaCpp(Xt, p = p, q = q, armamean = TRUE, inf_c = 1)
 BIC4_b <- -2 * fit4_b$loglik + ln * (p + q)
 
 p <- 0
 q <- 2
 fit5_a <- armaCpp(Xt, p = p, q = q, armamean = TRUE)
 fit5_b <- stats::arima(Xt, order = c(p, 0, q))
-BIC5_a <- BICarmaCpp(Xt, p = p, q = q, armamean = TRUE)
+BIC5_a <- BICarmaCpp(Xt, p = p, q = q, armamean = TRUE, inf_c = 1)
 BIC5_b <- -2 * fit5_b$loglik + ln * (p + q)
 
 p <- 2
 q <- 2
 fit6_a <- armaCpp(Xt, p = p, q = q, armamean = TRUE)
 fit6_b <- stats::arima(Xt, order = c(p, 0, q))
-BIC6_a <- BICarmaCpp(Xt, p = p, q = q, armamean = TRUE)
+BIC6_a <- BICarmaCpp(Xt, p = p, q = q, armamean = TRUE, inf_c = 1)
 BIC6_b <- -2 * fit6_b$loglik + ln * (p + q)
 
 ### Basic ARMA fitting with different orders using 'armaCpp'
@@ -94,7 +94,7 @@ for (p in 0:3) {
 pq_opt_b <- c(which(BIC_all_b == min(BIC_all_b), arr.ind = TRUE)) - 1
 pq_opt_b <- c(pq_opt_b[[1]], 0, pq_opt_b[[2]])
 pq_opt_a <- selectOrderBIC(Xt, pmin = 0, pmax = 3, qmin = 0, qmax = 3,
-                           armamean = TRUE)
+                           armamean = TRUE, inf_c = 1)
 
 test_that("ARMA model orders are selected correctly following the BIC", {
   
@@ -105,7 +105,7 @@ test_that("ARMA model orders are selected correctly following the BIC", {
 ### Optimal model fitting via 'armaoptCpp'
 
 mopt_b <- stats::arima(Xt, order = pq_opt_b)
-mopt_a <- armaoptCpp(Xt, 0, 3, 0, 3, TRUE)
+mopt_a <- armaoptCpp(Xt, 0, 3, 0, 3, TRUE, 1)
 
 test_that("Selected optimal model return the same fitting results", {